% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_path.R
\name{critical_path}
\alias{critical_path}
\title{Generate the critical path for a collection of related tasks}
\usage{
critical_path(data, gantt = F, start_date = Sys.Date())
}
\arguments{
\item{data}{A data frame of tasks with columns ID, name, duration, dependencies in that order.
Name of columns does not matter, only order.}

\item{gantt}{Boolean that specifies whether or not to produce a gantt chart of the results.}

\item{start_date}{Starting date for the project. Defaults to the current date.}
}
\value{
A list of results. First element is the id's of the critical path.
Second element is a data frame representation of the results that can be passed to the 'gantt' function.
Third element is the gantt chart if 'gantt' argument is true.
}
\description{
Generate the critical path for a collection of related tasks
}
\examples{
# Use provided sample data
data <- taskdata1

res <- critical_path(data, gantt = F)
}
