% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulations.R
\name{fetch_project_summaries_for_date_range}
\alias{fetch_project_summaries_for_date_range}
\title{Title}
\usage{
fetch_project_summaries_for_date_range(connection, start_date, end_date)
}
\arguments{
\item{end_date}{}
}
\description{
Title
}

