% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_path.R
\name{simulation}
\alias{simulation}
\title{Runs a simulation on project end time when certain tasks have uncertain durations}
\usage{
simulation(df, iter)
}
\arguments{
\item{df}{A data frame of tasks with columns ID, name, duration, id's of predecessrs (as a comma separated string),
minimum end time, most likely end time, and maximum end time (all in days) for the uncertain tasks in that order.
Rows without observations for the final three columns with not be treated as uncertain tasks.
Name of columns does not matter, only order. Type 'taskdatauncertain1' into the console for an example of valid data.}

\item{iter}{The number of times the simulation should run.}
}
\value{
A list of results.

\itemize{
\item \strong{durations} A vector of doubles (of the same size as 'iter') that contains the total project
duration for each itertation of the simulation.
\item \strong{histogram} A histogram of 'durations'
}
}
\description{
Runs a simulation on project end time when certain tasks have uncertain durations
}
\examples{
# Example using built in data
simulation(taskdatauncertain1, 10000)

}
