% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taskdata1}
\alias{taskdata1}
\title{Small collection of sample tasks}
\format{A data frame with 8 rows and 4 variables:
\describe{
  \item{id}{id for the task}
  \item{name}{name of the task}
  \item{duration}{duration of task, in days}
  \item{pred}{task id for the task's predeseccors}
}}
\usage{
taskdata1
}
\description{
Small collection of sample tasks
}
\keyword{datasets}
