% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taskdata2}
\alias{taskdata2}
\title{Larger collection of sample tasks with hanging tasks}
\format{A data frame with 14 rows and 4 variables:
\describe{
  \item{id}{id for the task}
  \item{name}{name of the task}
  \item{duration}{duration of task, in days}
  \item{pred}{id for the task's predecessors as a comma separated string}
}}
\usage{
taskdata2
}
\description{
Larger collection of sample tasks with hanging tasks
}
\keyword{datasets}
