% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taskdata3}
\alias{taskdata3}
\title{Collection of tasks with overlap for tasks in critical path.}
\format{A data frame with 13 rows and 4 variables:
\describe{
  \item{id}{id for the task}
  \item{name}{name of the task}
  \item{duration}{duration of task, in days}
  \item{pred}{id for the task's predecessors as a comma separated string}
}}
\usage{
taskdata3
}
\description{
Collection of tasks with overlap for tasks in critical path.
}
\keyword{datasets}
