% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taskdatauncertain1}
\alias{taskdatauncertain1}
\title{Sample task with uncertain durations}
\format{A data frame with 14 rows and 7 variables:
\describe{
  \item{id}{id for the task}
  \item{name}{name of the task}
  \item{duration}{duration of task, in days}
  \item{preds}{id for the task's predecessors as a comma separated string}
  \item{min}{minimum duration in days for the task}
  \item{mode}{most likely duration in days for the task}
  \item{max}{maximum duration in days for the task}
}}
\usage{
taskdatauncertain1
}
\description{
Sample task with uncertain durations
}
\keyword{datasets}
