% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taskdatauncertain2}
\alias{taskdatauncertain2}
\title{Larger collection of sample tasks with greater uncertainty in tasks}
\format{A data frame with 14 rows and 7 variables:
\describe{
  \item{id}{id for the task}
  \item{name}{name of the task}
  \item{duration}{duration of task, in days}
  \item{pred}{id for the task's predecessors as a comma separated string}
  \item{min}{minimum duration of task, in days}
  \item{mode}{most likely duration of task, in days}
  \item{max}{maximum duration of task, in days}
}}
\usage{
taskdatauncertain2
}
\description{
Larger collection of sample tasks with greater uncertainty in tasks
}
\keyword{datasets}
