% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulations.R
\name{fetch_aggregated_data_per_instrument_for_period}
\alias{fetch_aggregated_data_per_instrument_for_period}
\title{Title}
\usage{
fetch_aggregated_data_per_instrument_for_period(connection, start_date,
  end_date)
}
\arguments{
\item{end_date}{}
}
\description{
Title
}

