% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulations.R
\name{fetch_information_for_project}
\alias{fetch_information_for_project}
\title{Title}
\usage{
fetch_information_for_project(connection, project_name)
}
\arguments{
\item{project_name}{}
}
\description{
Title
}

