% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_path.R
\name{gantt}
\alias{gantt}
\title{Creates a Gantt chart of tasks in a project. This function is called by
'critical_path' if the 'gantt' argument is true.}
\usage{
gantt(df, raw = T)
}
\arguments{
\item{df}{A data frame of tasks. This data frame can either be raw data
(i.e not from the 'critical_path' function) or can be the data residing in
the 'results' element from the return value of the 'critical_path' function.
If the data is raw, if must have columns "ID, name, duration, dependencies"
in that order. These columns need not be named but they must be in that order.}

\item{raw}{Boolean indicating if the data is raw of if it has already been processed
by the 'critical_path' function.}
}
\value{
A gantt chart for the tasks. If raw is false, then this gantt chart will
color the critical path elements.
}
\description{
Creates a Gantt chart of tasks in a project. This function is called by
'critical_path' if the 'gantt' argument is true.
}
